refactor(content): consolidate content sources and rendering
Move content data and loaders under one content tree, keep blog island SSR beside the blog loader, and remove redundant wrappers. Make project skill guidance depend on responsibilities and current-tree discovery instead of brittle internal path lists.
