+++ .agents/skills/ryoppippi-com-content/SKILL.md
@@ -10,9 +10,9 @@ For discovering and updating external articles, podcasts, or YouTube appearances
 ## Blog posts
 
 - Blog posts live under `src/content/blog/<slug>/index.md`.
-- Read `src/content/blog/2024-10-12/index.md` for the supported Markdown extensions and frontmatter shape.
+- Read an existing post in the blog tree for the supported Markdown extensions and frontmatter shape.
 - Tweet embeds are explicit: `<Tweet id="1234567890" />`. Keep the original URL in an HTML comment; a URL alone does not create an embed.
-- Ox Content stores fetched Tweet records under `.cache/ox-content/twitter/` and self-hosted media under `static/ox-content/twitter/`. Do not hand-edit them; run `direnv exec . pnpm build` after changing an embed and commit the generated evidence it requires.
+- Ox Content's fetched Tweet records and self-hosted media are generated artifacts. Do not hand-edit them; run `direnv exec . pnpm build` after changing an embed and commit the generated evidence it requires.
 
 ## Translation and spelling
 
@@ -25,10 +25,10 @@ For discovering and updating external articles, podcasts, or YouTube appearances
 - Put components beside the post inside the blog tree. A supported island import is a capitalised default import from a relative `.tsx` path, such as `import Chart from './Chart.tsx'`, outside fenced code.
 - Use the import binding as the component tag, such as `<Chart />`. Props use Ox Content syntax: quoted strings, JSON in braces, or bare boolean attributes.
 - Resolved imports are removed from rendered Markdown; missing imports and unknown tags remain visible, so fix the path instead of hiding the error.
-- Read `src/content/islands.ts`, `src/content/island-renderer.ts`, and `src/content/markdown/render.ts` when the accepted syntax, path resolution, or SSR boundary is unclear.
+- Treat the current implementation and tests as the source of truth when the accepted syntax, path resolution, or SSR boundary is unclear.
 - Run `direnv exec . pnpm new` to create a post through `scripts/new-post.ts`; content rendering is part of the root Vite build.
 
 ## Portfolio data
 
-- Edit `src/contents/works/oss/list.json` for portfolio curation and follow `src/contents/works/oss/README.md`.
-- Do not hand-edit `src/contents/works/oss/stars.json`; refresh it with `pnpm update:oss-stars` when requested.
+- Edit `src/content/works/oss/list.json` for portfolio curation and follow `src/content/works/oss/README.md`.
+- Do not hand-edit `src/content/works/oss/stars.json`; refresh it with `pnpm update:oss-stars` when requested.
