--- src/content/article.css
@@ -1,2 +1,0 @@
-@source './blog/**/*.md';
-@source './blog/**/*.mdx';
