--- src/content/index.ts
@@ -1,4 +1,0 @@
-export type { ArticleMetadata, BlogPost, BlogPostMetadata } from './blog.ts';
-export type { ContentArtifact } from './artifact.ts';
-export type { IslandRenderer, MarkdownRenderer } from './markdown/render.ts';
-export type { ShowcaseProject } from './showcase.ts';
