--- src/content/paths.ts
@@ -1,11 +1,0 @@
-import path from 'node:path';
-
-const contentRoot = path.resolve(import.meta.dirname);
-
-export function blogDirectory(root = contentRoot) {
-	return path.join(root, 'blog');
-}
-
-export function showcaseDirectory(root = contentRoot) {
-	return path.join(root, 'showcase');
-}
