+++ src/content/works/oss/stars.json
