+++ src/pages/sponsors/index.ts
@@ -1,0 +1,26 @@
+import type { GeneratedFile } from '../index.ts';
+import type { SiteAssets } from '../../site/assets.ts';
+import { page, renderComponent } from '../../site/html.ts';
+import Sponsors from '../../site/templates/Sponsors.tsx';
+
+/**
+ * Renders the sponsors page.
+ *
+ * @param assets - Bundled site assets referenced by the page.
+ * @returns The generated sponsors page.
+ */
+export function sponsorsPage(assets: SiteAssets): GeneratedFile {
+	return {
+		path: 'sponsors/index.html',
+		sourcePaths: ['src/site/templates/Sponsors.tsx'],
+		content: page({
+			title: 'Sponsors',
+			pathname: '/sponsors/',
+			content: renderComponent(Sponsors, {}),
+			description:
+				"Support @ryoppippi's open-source projects, technical writing, and talks through GitHub Sponsors.",
+			assets,
+			style: 'sponsors',
+		}),
+	};
+}
