+++ src/site/feeds.ts
@@ -1,6 +1,5 @@
-import type { BlogPostMetadata } from '../content/index.ts';
+import type { BlogPostMetadata, PostListItem } from '../pages/blog/index.ts';
 import type { FeedChannelOptions, FeedItemInput, RenderedFeedFile } from '@ox-content/vite-plugin';
-import type { PostListItem } from './content.ts';
 import { renderFeedFiles, resolveFeedsOptions, writeFeedFiles } from '@ox-content/vite-plugin';
 import { SITE_COPYRIGHT, SITE_NAME, SITE_ORIGIN, SITE_SOCIAL_IMAGE_URL } from './consts.ts';
