+++ src/site/ox-content.ts
@@ -4,7 +4,7 @@ import {
 	type OxContentOptions,
 } from '@ox-content/vite-plugin';
 import { Route } from '../../routes.ts';
-import { OPEN_GRAPH_OPTIONS } from '../content/markdown/open-graph.ts';
+import { OPEN_GRAPH_OPTIONS } from '../pages/blog/markdown.ts';
 import { BLOG_FEED_OPTIONS } from './feeds.ts';
 
 export const BLOG_COLLECTION_PATTERNS = ['*.md', '*.mdx', '*/index.md', '*/index.mdx'] as const;
