--- src/styles/blog/article.css
@@ -1,6 +1,0 @@
-@layer theme, base, components, utilities;
-@import 'tailwindcss/utilities.css' layer(utilities) source(none);
-@import '../prose-theme.css';
-@reference '../tailwind.css';
-
-@import '../../content/article.css';
