+++ vite.config.ts
@@ -74,7 +74,10 @@ export default defineConfig(({ command, mode }) => ({
 			'.direnv/**',
 			'build/**',
 			'node_modules/**',
-			'src/contents/**',
+			'src/content/external-rss/*.json',
+			'src/content/works/oss/*.json',
+			'src/content/works/oss/README.md',
+			'src/content/publication.json',
 			'src/content/blog/**',
 			'src/content/showcase/**',
 			'static/**',
@@ -91,7 +94,10 @@ export default defineConfig(({ command, mode }) => ({
 			'.direnv/**',
 			'build/**',
 			'node_modules/**',
-			'src/contents/**',
+			'src/content/external-rss/*.json',
+			'src/content/works/oss/*.json',
+			'src/content/works/oss/README.md',
+			'src/content/publication.json',
 			'src/content/blog/**',
 			'src/content/showcase/**',
 			'static/**',
@@ -119,10 +125,9 @@ export default defineConfig(({ command, mode }) => ({
 					includeSource: [
 						'src/lib/**/*.ts',
 						'src/site/{assets,content-assets,dev-routes,dev-server}.ts',
-						'src/site/{generate,page-styles,pages,syntax-theme}.ts',
-						'src/content/{artifact,blog,island-renderer,islands,paths}.ts',
-						'src/content/blog/**/*.ts',
-						'src/content/markdown/**/*.ts',
+						'src/site/{generate,page-styles,syntax-theme}.ts',
+						'src/pages/*.ts',
+						'src/pages/**/*.ts',
 					],
 				},
 			},
